# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lehno",
    "version": "1",
    "description": "L'assistant des dates qui comptent. Le contrat est engendré depuis les schémas Zod de @lehno/contracts — il ne s'écrit pas à la main."
  },
  "servers": [
    {
      "url": "https://api.lehno.app/v1"
    },
    {
      "url": "http://localhost:3001/v1",
      "description": "développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/auth/otp": {
      "post": {
        "summary": "Demander un code de connexion à usage unique par courriel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "sent"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Défaillance du serveur — la requête n'y est pour rien, on peut réessayer sans la modifier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "4XX": {
            "description": "Refus — la requête ne satisfait pas le contrat (forme, valeur, droit) ; corriger avant de réessayer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "summary": "Vérifier le code reçu et ouvrir une session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  },
                  "deviceId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "referralCode": {
                    "type": "string",
                    "maxLength": 16
                  }
                },
                "required": [
                  "email",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "outcome": {
                          "type": "string",
                          "enum": [
                            "session"
                          ]
                        },
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "expiresIn": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "isNewAccount": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "outcome",
                        "accessToken",
                        "refreshToken",
                        "expiresIn",
                        "isNewAccount"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "outcome": {
                          "type": "string",
                          "enum": [
                            "registration"
                          ]
                        },
                        "registrationToken": {
                          "type": "string"
                        },
                        "expiresIn": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "deviceLimitReached": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "outcome",
                        "registrationToken",
                        "expiresIn",
                        "email",
                        "deviceLimitReached"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Défaillance du serveur — la requête n'y est pour rien, on peut réessayer sans la modifier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "4XX": {
            "description": "Refus — la requête ne satisfait pas le contrat (forme, valeur, droit) ; corriger avant de réessayer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Créer le compte : pseudo, appareil, et code de parrainage facultatif",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registrationToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 30,
                    "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._-]*$"
                  },
                  "deviceId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "referralCode": {
                    "type": "string",
                    "maxLength": 16
                  }
                },
                "required": [
                  "registrationToken",
                  "username",
                  "deviceId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "session"
                      ]
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0
                    },
                    "isNewAccount": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "signupCredits": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "referral": {
                      "type": "object",
                      "properties": {
                        "outcome": {
                          "type": "string",
                          "enum": [
                            "credited",
                            "unknown",
                            "self"
                          ]
                        },
                        "inviterUsername": {
                          "type": "string",
                          "nullable": true
                        },
                        "bonusCredits": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "outcome",
                        "inviterUsername",
                        "bonusCredits"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "outcome",
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "isNewAccount",
                    "signupCredits",
                    "referral"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Défaillance du serveur — la requête n'y est pour rien, on peut réessayer sans la modifier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "4XX": {
            "description": "Refus — la requête ne satisfait pas le contrat (forme, valeur, droit) ; corriger avant de réessayer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "summary": "Renouveler une session à partir du jeton de rafraîchissement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0
                    },
                    "isNewAccount": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "isNewAccount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Défaillance du serveur — la requête n'y est pour rien, on peut réessayer sans la modifier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "4XX": {
            "description": "Refus — la requête ne satisfait pas le contrat (forme, valeur, droit) ; corriger avant de réessayer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_failed",
                        "not_found",
                        "conflict",
                        "rate_limited",
                        "internal_error",
                        "unauthorized",
                        "forbidden",
                        "session_expired",
                        "refresh_reused",
                        "otp_invalid",
                        "otp_expired",
                        "otp_too_many_attempts",
                        "otp_rate_limited",
                        "reason_required",
                        "username_taken",
                        "username_invalid",
                        "device_limit_reached",
                        "account_suspended",
                        "account_pending_deletion",
                        "federated_token_invalid",
                        "federated_already_linked",
                        "email_disposable",
                        "waitlist_email_invalid",
                        "waitlist_rejected",
                        "contact_invalid",
                        "contact_rejected"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}